update: creating an abstract unit test class
